fix : main함수에 초기화 함수들만 존재
