+++ src/basic/ui/update/index.js
@@ -1,0 +1,6 @@
+export { updateCartUI } from "./updateCart";
+export { updateCartItemStyles } from "./updateCartItem";
+export { updateStockInfo } from "./updateStockInfo";
+export { updateBonusPoints } from "./updateBonusPoints";
+export { updateCartUIAfterCalculation } from "./updateCartUIAfterCalculation";
+export { updateCartItemPrices } from "./updateCartItemPrices";
